Refactored sign up screen
