+++ src/api/tracker.js
@@ -2,5 +2,5 @@ import axios from 'axios';
 
 //base url from ngrok, must get new URL every 8 hours
 export default axios.create({
-  baseURL:'http://3b55bf78.ngrok.io'
+  baseURL:'http://03359d5f.ngrok.io'
 });
