+++ src/screens/SignupScreen.js
@@ -1,25 +1,15 @@
 import React, {useState} from 'react';
 import {View,StyleSheet, TextInput,TouchableOpacity} from 'react-native';
 import{Text,Input,Button} from 'react-native-elements';
-import {connect} from 'react-redux';
+import AuthForm from '../components/AuthForm'
 import {signUp} from '../actions';
-import Spacer from '../components/Spacer'
+import {connect} from 'react-redux';
 
-const SignupScreen= ({navigation,Auth,signUp},props)=>{
-  const [email, setEmail]=useState('');
-  const [password,setPassword]= useState('');
+const SignupScreen= ({navigation,Auth,signUp})=>{
+
 
   return <View style={styles.container}>
-    <Spacer>
-      <Text h3>Sign Up For Tracker</Text>
-    </Spacer>
-    <Input label ="Email" value={email} onChangeText={setEmail} autoCapitalize="none" autoCorrect ={false}/>
-    <Spacer />
-    <Input  secureTextEntry label="Password" value={password} onChangeText={setPassword} autoCapitalize="none" autoCorrect ={false}/>
-    {Auth.errorMessage ? <Text style= {styles.errorMessage}>{Auth.errorMessage}</Text> : null}
-    <Spacer>
-      <Button title = "Sign Up" onPress={()=>signUp({email: email,password:password})}/>
-    </Spacer>
+  <AuthForm headerText="test" buttonText="test" onSubmit={signUp}/>
     </View>
 };
 
@@ -35,18 +25,7 @@ const styles = StyleSheet.create({
     justifyContent:'center',
     marginBottom: 200
   },
-  errorMessage:{
-    fontSize: 16,
-    color:'red',
-    marginLeft:15,
-    marginTop:15
-  }
 });
 
-const mapStateToProps=(state)=>{
-  return {
-    Auth: state.Auth
-  }
-}
 
-export default connect(mapStateToProps,{signUp})(SignupScreen);
+export default connect(null,{signUp})(SignupScreen);
